+++ asdf/tests/test_fits_embed.py
@@ -319,15 +319,17 @@ def test_version_mismatch_file():
                 ignore_version_mismatch=False) as fits_handle:
             assert fits_handle.tree['a'] == complex(0j)
     # This is the warning that we expect from opening the FITS file
-    assert len(w) == 2, display_warnings(w)
-    assert str(w[0].message) == (
-        "'tag:stsci.edu:asdf/core/complex' with version 7.0.0 found in file "
-        "'{}', but latest supported version is 1.0.0".format(testfile)
-    )
-    assert str(w[1].message) == (
-        "'tag:stsci.edu:asdf/core/asdf' with version 1.0.0 found in file "
-        "'{}', but latest supported version is 1.1.0".format(testfile)
-    )
+    expected_messages = {
+        (
+            "'tag:stsci.edu:asdf/core/complex' with version 7.0.0 found in file "
+            "'{}', but latest supported version is 1.0.0".format(testfile)
+        ),
+        (
+            "'tag:stsci.edu:asdf/core/asdf' with version 1.0.0 found in file "
+            "'{}', but latest supported version is 1.1.0".format(testfile)
+        ),
+    }
+    assert expected_messages == {warn.message.args[0] for warn in w}, display_warnings(w)
 
     # Make sure warning does not occur when warning is ignored (default)
     with pytest.warns(None) as w:
@@ -339,15 +341,17 @@ def test_version_mismatch_file():
         with fits_embed.AsdfInFits.open(testfile,
                 ignore_version_mismatch=False) as fits_handle:
             assert fits_handle.tree['a'] == complex(0j)
-    assert len(w) == 2
-    assert str(w[0].message) == (
-        "'tag:stsci.edu:asdf/core/complex' with version 7.0.0 found in file "
-        "'{}', but latest supported version is 1.0.0".format(testfile)
-    )
-    assert str(w[1].message) == (
-        "'tag:stsci.edu:asdf/core/asdf' with version 1.0.0 found in file "
-        "'{}', but latest supported version is 1.1.0".format(testfile)
-    )
+    expected_messages = {
+        (
+            "'tag:stsci.edu:asdf/core/complex' with version 7.0.0 found in file "
+            "'{}', but latest supported version is 1.0.0".format(testfile)
+        ),
+        (
+            "'tag:stsci.edu:asdf/core/asdf' with version 1.0.0 found in file "
+            "'{}', but latest supported version is 1.1.0".format(testfile)
+        ),
+    }
+    assert expected_messages == {warn.message.args[0] for warn in w}, display_warnings(w)
 
     # Make sure warning does not occur when warning is ignored (default)
     with pytest.warns(None) as w: